# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dgeneris Bid & Tender Solutions API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WebhookAck {
  received: boolean;
}

export interface Enquiry {
  id: number;
  name: string;
  companyName: string;
  email: string;
  phone: string;
  /** care or cleaning */
  sector: string;
  serviceRequired: string;
  /** @nullable */
  tenderDeadline?: string | null;
  /** @nullable */
  estimatedBudget?: string | null;
  message: string;
  /** new, in_review, responded, closed */
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  newsletterOptIn: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface EnquiryInput {
  name: string;
  companyName: string;
  email: string;
  phone: string;
  sector: string;
  serviceRequired: string;
  /** @nullable */
  tenderDeadline?: string | null;
  /** @nullable */
  estimatedBudget?: string | null;
  message: string;
  newsletterOptIn?: boolean;
}

export interface EnquiryUpdate {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  adminNotes?: string | null;
}

export interface Faq {
  id: number;
  question: string;
  answer: string;
  /** @nullable */
  category?: string | null;
  sortOrder: number;
  published: boolean;
  createdAt: string;
}

export interface FaqInput {
  question: string;
  answer: string;
  /** @nullable */
  category?: string | null;
  sortOrder: number;
  published?: boolean;
}

export interface FaqUpdate {
  /** @nullable */
  question?: string | null;
  /** @nullable */
  answer?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  published?: boolean | null;
}

export interface Testimonial {
  id: number;
  clientName: string;
  companyName: string;
  sector: string;
  quote: string;
  /** @nullable */
  rating?: number | null;
  published: boolean;
  createdAt: string;
}

export interface TestimonialInput {
  clientName: string;
  companyName: string;
  sector: string;
  quote: string;
  /** @nullable */
  rating?: number | null;
  published?: boolean;
}

export interface TestimonialUpdate {
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  quote?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  published?: boolean | null;
}

export interface CaseStudy {
  id: number;
  clientSector: string;
  tenderType: string;
  objective: string;
  challenge: string;
  supportProvided: string;
  outcome: string;
  /** @nullable */
  clientFeedback?: string | null;
  published: boolean;
  /** Marks this as a placeholder example, not a real case */
  isExample: boolean;
  createdAt: string;
}

export interface CaseStudyInput {
  clientSector: string;
  tenderType: string;
  objective: string;
  challenge: string;
  supportProvided: string;
  outcome: string;
  /** @nullable */
  clientFeedback?: string | null;
  published?: boolean;
  isExample?: boolean;
}

export interface CaseStudyUpdate {
  /** @nullable */
  clientSector?: string | null;
  /** @nullable */
  tenderType?: string | null;
  /** @nullable */
  objective?: string | null;
  /** @nullable */
  challenge?: string | null;
  /** @nullable */
  supportProvided?: string | null;
  /** @nullable */
  outcome?: string | null;
  /** @nullable */
  clientFeedback?: string | null;
  /** @nullable */
  published?: boolean | null;
  /** @nullable */
  isExample?: boolean | null;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  category?: string | null;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  category?: string | null;
  published?: boolean;
}

export interface BlogPostUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  published?: boolean | null;
}

export interface NewsletterInput {
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface NewsletterSubscriber {
  id: number;
  email: string;
  /** @nullable */
  name?: string | null;
  subscribedAt: string;
}

export interface CheckoutSessionInput {
  /** tender_research | bid_review | full_tender | retainer | custom */
  packageType: string;
  clientName: string;
  clientEmail: string;
  companyName: string;
  description: string;
  /**
     * Amount in pence (GBP). Null for custom invoice
     * @nullable
     */
  amount?: number | null;
}

export interface CheckoutSessionResult {
  url: string;
  sessionId: string;
}

export interface Order {
  id: number;
  clientName: string;
  clientEmail: string;
  companyName: string;
  packageType: string;
  description: string;
  amountPence: number;
  /** pending | paid | failed | refunded */
  status: string;
  /** @nullable */
  stripeSessionId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Project {
  id: number;
  clientName: string;
  clientEmail: string;
  companyName: string;
  serviceType: string;
  /** @nullable */
  tenderDeadline?: string | null;
  /** enquiry_received | awaiting_documents | under_review | information_required | drafting | client_review | revision | final_quality_check | ready_for_submission | completed */
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  /** @nullable */
  orderId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectUpdate {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  /** @nullable */
  tenderDeadline?: string | null;
}

export interface ProjectMessage {
  id: number;
  projectId: number;
  /** client | admin */
  senderType: string;
  senderName: string;
  message: string;
  createdAt: string;
}

export interface ProjectMessageInput {
  senderType: string;
  senderName: string;
  message: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface AdminStats {
  totalEnquiries: number;
  newEnquiries: number;
  totalOrders: number;
  totalRevenuePence: number;
  activeProjects: number;
  completedProjects: number;
  totalTestimonials: number;
  totalBlogPosts: number;
  totalSubscribers: number;
  recentEnquiries: Enquiry[];
  recentOrders: Order[];
  projectsByStatus: StatusCount[];
}

export type ListBlogPostsParams = {
/**
 * Filter by category
 * @nullable
 */
category?: string | null;
};

export type ListAdminEnquiriesParams = {
/**
 * @nullable
 */
status?: string | null;
};