# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Dgeneris Bid & Tender Solutions API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by category
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blog-posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Public: Newsletter ───────────────────────────────────────────────────
...
  /checkout/create-session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: Stripe checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/projects/{id}:
    patch:
      operationId: updateAdminProject
      tags: [admin]
      summary: Update a project (status, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /admin/blog-posts:
    get:
      operationId: listAdminBlogPosts
      tags: [admin]
      summary: List all blog posts (admin, including drafts)
      responses:
        "200":
          description: All blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [admin]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /admin/blog-posts/{id}:
    patch:
      operationId: updateBlogPost
      tags: [admin]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [admin]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /admin/case-studies:
    post:
      operationId: createCaseStudy
      tags: [admin]
      summary: Create a case study
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyInput"
      responses:
        "201":
          description: Created case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
  /admin/case-studies/{id}:
    patch:
      operationId: updateCaseStudy
      tags: [admin]
      summary: Update a case study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyUpdate"
      responses:
        "200":
          description: Updated case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
    delete:
      operationId: deleteCaseStudy
      tags: [admin]
      summary: Delete a case study
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    CaseStudy:
      type: object
      properties:
        id:
          type: number
        clientSector:
          type: string
        tenderType:
          type: string
        objective:
          type: string
        challenge:
          type: string
        supportProvided:
          type: string
        outcome:
          type: string
        clientFeedback:
          type: ["string", "null"]
        published:
          type: boolean
        isExample:
          type: boolean
          description: "Marks this as a placeholder example, not a real case"
        createdAt:
          type: string
      required: [id, clientSector, tenderType, objective, challenge, supportProvided, outcome, published, isExample, createdAt]
    CaseStudyInput:
      type: object
      properties:
        clientSector:
          type: string
        tenderType:
          type: string
        objective:
          type: string
        challenge:
          type: string
        supportProvided:
          type: string
        outcome:
          type: string
        clientFeedback:
          type: ["string", "null"]
        published:
          type: boolean
        isExample:
          type: boolean
      required: [clientSector, tenderType, objective, challenge, supportProvided, outcome]
    CaseStudyUpdate:
      type: object
      properties:
        clientSector:
          type: ["string", "null"]
        tenderType:
          type: ["string", "null"]
        objective:
          type: ["string", "null"]
        challenge:
          type: ["string", "null"]
        supportProvided:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
        clientFeedback:
          type: ["string", "null"]
        published:
          type: ["boolean", "null"]
        isExample:
          type: ["boolean", "null"]

    # ── BlogPost ─────────────────────────────────────────────────────────────
    BlogPost:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, slug, excerpt, content, published, createdAt, updatedAt]
    BlogPostInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
        published:
          type: boolean
      required: [title, slug, excerpt, content]
    BlogPostUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        slug:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        published:
          type: ["boolean", "null"]

    # ── Newsletter ──────────────────────────────────────────────────────────
...
    CheckoutSessionInput:
      type: object
      properties:
        packageType:
          type: string
          description: "tender_research | bid_review | full_tender | retainer | custom"
        clientName:
          type: string
        clientEmail:
          type: string
        companyName:
          type: string
        description:
          type: string
        amount:
          type: ["number", "null"]
          description: "Amount in pence (GBP). Null for custom invoice"
      required: [packageType, clientName, clientEmail, companyName, description]
    CheckoutSessionResult:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required: [url, sessionId]

    # ── Order ────────────────────────────────────────────────────────────────
...
    Project:
      type: object
      properties:
        id:
          type: number
        clientName:
          type: string
        clientEmail:
          type: string
        companyName:
          type: string
        serviceType:
          type: string
        tenderDeadline:
          type: ["string", "null"]
        status:
          type: string
          description: "enquiry_received | awaiting_documents | under_review | information_required | drafting | client_review | revision | final_quality_check | ready_for_submission | completed"
        adminNotes:
          type: ["string", "null"]
        orderId:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, clientName, clientEmail, companyName, serviceType, status, createdAt, updatedAt]
    ProjectUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        tenderDeadline:
          type: ["string", "null"]

    # ── Project Messages ────────────────────────────────────────────────────
...