# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Dgeneris Bid & Tender Solutions API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  category?: string | null;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}